Bug fix for edge case when creating a new Post
This check was added to account for an edge case where date_said == custom_age,
and kid_id is blank.  Params with `custom_age` triggers the elsif block.
This blank kid_id was causing the elsif block to error out
on `Kid.find(params[:post][:kid_id]).birthdate`, as kid_id was blank.
Consequently, the @post.save line in :create was never being reached,
so model validation was never being triggered.
Thusly, this `if` check is now accounting for this edge case,
to ensure that the @post.save line is reached in the :create method,
thereby allowing validation in the model to be checked on kid_id presense.